refac: 구매 객체 생성
